requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosqlite==0.19.0
fastapi==0.109.0
Jinja2==3.1.6
pydantic==2.5.3
python-dotenv==1.0.1
python-multipart==0.0.6
SQLAlchemy==2.0.38
uvicorn[standard]==0.27.0
